requirements.txt

fastmcp
pydantic
httpx
python-dotenv